+++ tests/contrib/grpc/test_grpc.py
@@ -2,7 +2,6 @@ import threading
 import time
 
 import grpc
-from grpc._grpcio_metadata import __version__ as _GRPC_VERSION
 from grpc.framework.foundation import logging_pool
 import pytest
 import six
@@ -19,96 +18,16 @@ from ddtrace.contrib.grpc import unpatch
 from ddtrace.contrib.grpc.patch import _unpatch_server
 from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
 from tests.utils import TracerTestCase
-from tests.utils import override_env
 from tests.utils import snapshot
 
-from .hello_pb2 import HelloReply
+from .common import _GRPC_PORT
+from .common import _GRPC_VERSION
+from .common import GrpcBaseTestCase
 from .hello_pb2 import HelloRequest
-from .hello_pb2_grpc import HelloServicer
 from .hello_pb2_grpc import HelloStub
-from .hello_pb2_grpc import add_HelloServicer_to_server
 
 
-_GRPC_PORT = 50531
-_GRPC_VERSION = tuple([int(i) for i in _GRPC_VERSION.split(".")])
-
-
-def _check_test_range(value):
-    from ddtrace.appsec._iast._taint_tracking import get_tainted_ranges
-
-    ranges = get_tainted_ranges(value)
-    assert len(ranges) == 1
-    source = ranges[0].source
-    assert source.name == "http.request.grpc_body"
-    assert hasattr(source, "value")
-
-
-class GrpcTestCase(TracerTestCase):
-    def setUp(self):
-        super(GrpcTestCase, self).setUp()
-        patch()
-        Pin.override(constants.GRPC_PIN_MODULE_SERVER, tracer=self.tracer)
-        Pin.override(constants.GRPC_PIN_MODULE_CLIENT, tracer=self.tracer)
-        self._start_server()
-
-    @TracerTestCase.run_in_subprocess(env_overrides=dict(DD_IAST_ENABLED="1"))
-    def test_taint_iast_single(self):
-        with override_env({"DD_IAST_ENABLED": "True"}):
-            with self.override_config("grpc", dict(service_name="myclientsvc")):
-                with self.override_config("grpc_server", dict(service_name="myserversvc")):
-                    channel1 = grpc.insecure_channel("localhost:%d" % (_GRPC_PORT))
-                    stub1 = HelloStub(channel1)
-                    res = stub1.SayHello(HelloRequest(name="test"))
-                    assert "response" in res
-                    assert hasattr(res["response"], "value")
-                    assert hasattr(res["response"].value, "message")
-                    _check_test_range(res["response"].value.message)
-
-    @TracerTestCase.run_in_subprocess(env_overrides=dict(DD_IAST_ENABLED="1"))
-    def test_taint_iast_twice(self):
-        with override_env({"DD_IAST_ENABLED": "True"}):
-            with self.override_config("grpc", dict(service_name="myclientsvc")):
-                with self.override_config("grpc_server", dict(service_name="myserversvc")):
-                    channel1 = grpc.insecure_channel("localhost:%d" % (_GRPC_PORT))
-                    stub1 = HelloStub(channel1)
-                    responses_iterator = stub1.SayHelloTwice(HelloRequest(name="test"))
-                    for res in responses_iterator:
-                        assert "response" in res
-                        assert hasattr(res["response"], "value")
-                        assert hasattr(res["response"].value, "message")
-                        _check_test_range(res["response"].value.message)
-
-    @TracerTestCase.run_in_subprocess(env_overrides=dict(DD_IAST_ENABLED="1"))
-    def test_taint_iast_repeatedly(self):
-        with override_env({"DD_IAST_ENABLED": "True"}):
-            with self.override_config("grpc", dict(service_name="myclientsvc")):
-                with self.override_config("grpc_server", dict(service_name="myserversvc")):
-                    channel1 = grpc.insecure_channel("localhost:%d" % (_GRPC_PORT))
-                    stub1 = HelloStub(channel1)
-                    requests_iterator = iter(
-                        HelloRequest(name=name) for name in ["first", "second", "third", "fourth", "fifth"]
-                    )
-                    responses_iterator = stub1.SayHelloRepeatedly(requests_iterator)
-                    for res in responses_iterator:
-                        assert "response" in res
-                        assert hasattr(res["response"], "value")
-                        assert hasattr(res["response"].value, "message")
-                        _check_test_range(res["response"].value.message)
-
-    @TracerTestCase.run_in_subprocess(env_overrides=dict(DD_IAST_ENABLED="1"))
-    def test_taint_iast_last(self):
-        with override_env({"DD_IAST_ENABLED": "True"}):
-            with self.override_config("grpc", dict(service_name="myclientsvc")):
-                with self.override_config("grpc_server", dict(service_name="myserversvc")):
-                    channel1 = grpc.insecure_channel("localhost:%d" % (_GRPC_PORT))
-                    stub1 = HelloStub(channel1)
-                    requests_iterator = iter(HelloRequest(name=name) for name in ["first", "second"])
-                    res = stub1.SayHelloLast(requests_iterator)
-                    assert "response" in res
-                    assert hasattr(res["response"], "value")
-                    assert hasattr(res["response"].value, "message")
-                    _check_test_range(res["response"].value.message)
-
+class GrpcTestCase(GrpcBaseTestCase):
     @TracerTestCase.run_in_subprocess(env_overrides=dict(DD_SERVICE="mysvc", DD_TRACE_SPAN_ATTRIBUTE_SCHEMA="v1"))
     def test_user_specified_service_v1(self):
         with grpc.insecure_channel("localhost:%d" % (_GRPC_PORT)) as channel:
@@ -196,14 +115,6 @@ class GrpcTestCase(TracerTestCase):
         client_span, server_span = spans
         assert client_span.service == "mygrpc"
 
-    def tearDown(self):
-        self._stop_server()
-        # Remove any remaining spans
-        self.tracer.pop()
-        # Unpatch grpc
-        unpatch()
-        super(GrpcTestCase, self).tearDown()
-
     def get_spans_with_sync_and_assert(self, size=0, retry=20):
         # testing instrumentation with grpcio < 1.14.0 presents a problem for
         # checking spans written to the dummy tracer
@@ -222,17 +133,6 @@ class GrpcTestCase(TracerTestCase):
             time.sleep(0.1)
 
         return spans
-
-    def _start_server(self):
-        self._server_pool = logging_pool.pool(1)
-        self._server = grpc.server(self._server_pool)
-        self._server.add_insecure_port("[::]:%d" % (_GRPC_PORT))
-        add_HelloServicer_to_server(_HelloServicer(), self._server)
-        self._server.start()
-
-    def _stop_server(self):
-        self._server.stop(None)
-        self._server_pool.shutdown(wait=True)
 
     def _check_client_span(self, span, service, method_name, method_kind):
         self.assert_is_measured(span)
@@ -748,66 +648,6 @@ class GrpcTestCase(TracerTestCase):
         client_span, server_span = spans
         assert client_span.name == "grpc.client.request", "Expected 'grpc.client.request', got %s" % client_span.name
         assert server_span.name == "grpc.server.request", "Expected 'grpc.server.request', got %s" % server_span.name
-
-
-class _HelloServicer(HelloServicer):
-    def SayHello(self, request, context):
-        if request.name == "propogator":
-            metadata = context.invocation_metadata()
-            context.set_code(grpc.StatusCode.OK)
-            message = ";".join(w.key + "=" + w.value for w in metadata if w.key.startswith("x-datadog"))
-            return HelloReply(message=message)
-
-        if request.name == "abort":
-            context.abort(grpc.StatusCode.ABORTED, "aborted")
-
-        if request.name == "exception":
-            context.abort(grpc.StatusCode.INVALID_ARGUMENT, "exception")
-
-        return HelloReply(message="Hello {}".format(request.name))
-
-    def SayHelloTwice(self, request, context):
-        yield HelloReply(message="first response")
-
-        if request.name == "exception":
-            context.abort(grpc.StatusCode.RESOURCE_EXHAUSTED, "exception")
-
-        if request.name == "once":
-            # Mimic behavior of scenario where only one result is expected from
-            # streaming response and the RPC is successfully terminated, as is
-            # the case with grpc_helpers._StreamingResponseIterator in the
-            # Google API Library wraps a _MultiThreadedRendezvous future. An
-            # example of this iterator only called once is in the Google Cloud
-            # Firestore library.
-            # https://github.com/googleapis/python-api-core/blob/f87bccbfda11d2c2d1a2ddb6611c1209e29289d9/google/api_core/grpc_helpers.py#L80-L116
-            # https://github.com/googleapis/python-firestore/blob/e57258c51e4b4aa664cc927454056412756fc7ac/google/cloud/firestore_v1/document.py#L400-L404
-            return
-
-        yield HelloReply(message="secondresponse")
-
-    def SayHelloLast(self, request_iterator, context):
-        names = [r.name for r in list(request_iterator)]
-
-        if "exception" in names:
-            context.abort(grpc.StatusCode.INVALID_ARGUMENT, "exception")
-
-        return HelloReply(message="{}".format(";".join(names)))
-
-    def SayHelloRepeatedly(self, request_iterator, context):
-        last_request = None
-        for request in request_iterator:
-            if last_request is not None:
-                yield HelloReply(message="{}".format(";".join([last_request.name, request.name])))
-                last_request = None
-            else:
-                last_request = request
-
-        # response for dangling request
-        if last_request is not None:
-            yield HelloReply(message="{}".format(last_request.name))
-
-    def SayHelloUnknown(self, request, context):
-        yield HelloReply(message="unknown")
 
 
 class _CustomException(Exception):
